+++ x-pack/solutions/observability/plugins/slo/server/services/management/compute_slo_health.ts
@@ -41,7 +41,7 @@ interface SummaryAggResults {
 
 const BATCH_SIZE = 100;
 
-export class ComputeHealth {
+export class ComputeSLOHealth {
   constructor(
     private readonly esClient: ElasticsearchClient,
     private readonly soClient: SavedObjectsClient,
